# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  active: "active",
  archived: "archived",
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
}

export type UpdateProjectBodyStatus =
  (typeof UpdateProjectBodyStatus)[keyof typeof UpdateProjectBodyStatus];

export const UpdateProjectBodyStatus = {
  active: "active",
  archived: "archived",
} as const;

export interface UpdateProjectBody {
  name?: string;
  /** @nullable */
  description?: string | null;
  status?: UpdateProjectBodyStatus;
}

export interface ProjectSummary {
  projectId: number;
  sessionCount: number;
  committedCount: number;
  parkedCount: number;
  /** @nullable */
  recentMode?: string | null;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  active: "active",
  closed: "closed",
} as const;

export interface Session {
  id: number;
  projectId: number;
  title: string;
  /** @nullable */
  mode?: string | null;
  status: SessionStatus;
  messageCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateSessionBody {
  title: string;
  /** @nullable */
  mode?: string | null;
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

/**
 * @nullable
 */
export type MessageCatchPayload = { [key: string]: unknown } | null;

export interface Message {
  id: number;
  sessionId: number;
  role: MessageRole;
  content: string;
  /** @nullable */
  intentType?: string | null;
  /** @nullable */
  catchPayload?: MessageCatchPayload;
  createdAt: string;
}

export interface SessionWithMessages {
  session: Session;
  messages: Message[];
}

export type EntryStatus = (typeof EntryStatus)[keyof typeof EntryStatus];

export const EntryStatus = {
  committed: "committed",
  parked: "parked",
  draft: "draft",
  archived: "archived",
} as const;

export type EntrySeverity = (typeof EntrySeverity)[keyof typeof EntrySeverity];

export const EntrySeverity = {
  blocker: "blocker",
  parked: "parked",
  committed: "committed",
  neutral: "neutral",
} as const;

export interface Entry {
  id: number;
  projectId: number;
  /** @nullable */
  sessionId?: number | null;
  status: EntryStatus;
  title: string;
  /** @nullable */
  summary?: string | null;
  severity: EntrySeverity;
  /** @nullable */
  verb?: string | null;
  isViolation: boolean;
  deviation: boolean;
  /** @nullable */
  mode?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateEntryBodyStatus =
  (typeof CreateEntryBodyStatus)[keyof typeof CreateEntryBodyStatus];

export const CreateEntryBodyStatus = {
  committed: "committed",
  parked: "parked",
  draft: "draft",
} as const;

export type CreateEntryBodySeverity =
  (typeof CreateEntryBodySeverity)[keyof typeof CreateEntryBodySeverity];

export const CreateEntryBodySeverity = {
  blocker: "blocker",
  parked: "parked",
  committed: "committed",
  neutral: "neutral",
} as const;

export interface CreateEntryBody {
  /** @nullable */
  sessionId?: number | null;
  status: CreateEntryBodyStatus;
  title: string;
  /** @nullable */
  summary?: string | null;
  severity: CreateEntryBodySeverity;
  /** @nullable */
  verb?: string | null;
  /** @nullable */
  mode?: string | null;
}

export type UpdateEntryBodyStatus =
  (typeof UpdateEntryBodyStatus)[keyof typeof UpdateEntryBodyStatus];

export const UpdateEntryBodyStatus = {
  committed: "committed",
  parked: "parked",
  draft: "draft",
  archived: "archived",
} as const;

export type UpdateEntryBodySeverity =
  (typeof UpdateEntryBodySeverity)[keyof typeof UpdateEntryBodySeverity];

export const UpdateEntryBodySeverity = {
  blocker: "blocker",
  parked: "parked",
  committed: "committed",
  neutral: "neutral",
} as const;

export interface UpdateEntryBody {
  status?: UpdateEntryBodyStatus;
  title?: string;
  /** @nullable */
  summary?: string | null;
  severity?: UpdateEntryBodySeverity;
}

export type ChatHistoryItemRole =
  (typeof ChatHistoryItemRole)[keyof typeof ChatHistoryItemRole];

export const ChatHistoryItemRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatHistoryItem {
  role: ChatHistoryItemRole;
  content: string;
}

export interface EntryContext {
  id: number;
  title: string;
  status: string;
}

export interface ChatRequest {
  sessionId: number;
  projectId: number;
  message: string;
  mode: string;
  history?: ChatHistoryItem[];
  entries?: EntryContext[];
}

/**
 * @nullable
 */
export type ChatResponseCatchPayload = { [key: string]: unknown } | null;

export interface ChatResponse {
  content: string;
  /** @nullable */
  intentType?: string | null;
  /** @nullable */
  catchPayload?: ChatResponseCatchPayload;
  messageId: number;
}

export type ListEntriesParams = {
  status?: ListEntriesStatus;
};

export type ListEntriesStatus =
  (typeof ListEntriesStatus)[keyof typeof ListEntriesStatus];

export const ListEntriesStatus = {
  committed: "committed",
  parked: "parked",
  draft: "draft",
  archived: "archived",
} as const;