# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Atlas - Strategic Thinking Partner API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management
  - name: sessions
    description: Chat sessions
  - name: messages
    description: Chat messages
  - name: entries
    description: Decision ledger entries
  - name: chat
    description: AI chat
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/summary:
    get:
      operationId: getProjectSummary
      tags: [projects]
      summary: Get project summary stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectSummary"

  /projects/{projectId}/sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List sessions for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a new session
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a session with its messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionWithMessages"
        "404":
          description: Not found
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /sessions/{sessionId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a session
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

  /projects/{projectId}/entries:
    get:
      operationId: listEntries
      tags: [entries]
      summary: List decision ledger entries for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [committed, parked, draft, archived]
      responses:
        "200":
          description: List of entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
    post:
      operationId: createEntry
      tags: [entries]
      summary: Create a ledger entry
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEntryBody"
      responses:
        "201":
          description: Created entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"

  /entries/{id}:
    patch:
      operationId: updateEntry
      tags: [entries]
      summary: Update a ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEntryBody"
      responses:
        "200":
          description: Updated entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Entry"
        "404":
          description: Not found
    delete:
      operationId: deleteEntry
      tags: [entries]
      summary: Delete a ledger entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /chat:
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get an AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - createdAt
        - updatedAt

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name

    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, archived]

    ProjectSummary:
      type: object
      properties:
        projectId:
          type: integer
        sessionCount:
          type: integer
        committedCount:
          type: integer
        parkedCount:
          type: integer
        recentMode:
          type: ["string", "null"]
      required:
        - projectId
        - sessionCount
        - committedCount
        - parkedCount

    Session:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        mode:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, closed]
        messageCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - title
        - status
        - messageCount
        - createdAt
        - updatedAt

    CreateSessionBody:
      type: object
      properties:
        title:
          type: string
        mode:
          type: ["string", "null"]
      required:
        - title

    SessionWithMessages:
      type: object
      properties:
        session:
          $ref: "#/components/schemas/Session"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required:
        - session
        - messages

    Message:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        intentType:
          type: ["string", "null"]
        catchPayload:
          type: ["object", "null"]
        createdAt:
          type: string
      required:
        - id
        - sessionId
        - role
        - content
        - createdAt

    Entry:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        sessionId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [committed, parked, draft, archived]
        title:
          type: string
        summary:
          type: ["string", "null"]
        severity:
          type: string
          enum: [blocker, parked, committed, neutral]
        verb:
          type: ["string", "null"]
        isViolation:
          type: boolean
        deviation:
          type: boolean
        mode:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectId
        - status
        - title
        - severity
        - isViolation
        - deviation
        - createdAt
        - updatedAt

    CreateEntryBody:
      type: object
      properties:
        sessionId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [committed, parked, draft]
        title:
          type: string
        summary:
          type: ["string", "null"]
        severity:
          type: string
          enum: [blocker, parked, committed, neutral]
        verb:
          type: ["string", "null"]
        mode:
          type: ["string", "null"]
      required:
        - title
        - status
        - severity

    UpdateEntryBody:
      type: object
      properties:
        status:
          type: string
          enum: [committed, parked, draft, archived]
        title:
          type: string
        summary:
          type: ["string", "null"]
        severity:
          type: string
          enum: [blocker, parked, committed, neutral]

    ChatRequest:
      type: object
      properties:
        sessionId:
          type: integer
        projectId:
          type: integer
        message:
          type: string
        mode:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatHistoryItem"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/EntryContext"
      required:
        - sessionId
        - projectId
        - message
        - mode

    ChatHistoryItem:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content

    EntryContext:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        status:
          type: string
      required:
        - id
        - title
        - status

    ChatResponse:
      type: object
      properties:
        content:
          type: string
        intentType:
          type: ["string", "null"]
        catchPayload:
          type: ["object", "null"]
        messageId:
          type: integer
      required:
        - content
        - messageId