# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "archived"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
});

/**
 * @summary Get a project
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "archived"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "archived"]).optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  status: zod.enum(["active", "archived"]),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get project summary stats
 */
export const GetProjectSummaryParams = zod.object({
  id: zod.coerce.number(),
});

export const GetProjectSummaryResponse = zod.object({
  projectId: zod.number(),
  sessionCount: zod.number(),
  committedCount: zod.number(),
  parkedCount: zod.number(),
  recentMode: zod.string().nullish(),
});

/**
 * @summary List sessions for a project
 */
export const ListSessionsParams = zod.object({
  projectId: zod.coerce.number(),
});

export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  title: zod.string(),
  mode: zod.string().nullish(),
  status: zod.enum(["active", "closed"]),
  messageCount: zod.number(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Create a new session
 */
export const CreateSessionParams = zod.object({
  projectId: zod.coerce.number(),
});

export const CreateSessionBody = zod.object({
  title: zod.string(),
  mode: zod.string().nullish(),
});

/**
 * @summary Get a session with its messages
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  session: zod.object({
    id: zod.number(),
    projectId: zod.number(),
    title: zod.string(),
    mode: zod.string().nullish(),
    status: zod.enum(["active", "closed"]),
    messageCount: zod.number(),
    createdAt: zod.string(),
    updatedAt: zod.string(),
  }),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      sessionId: zod.number(),
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
      intentType: zod.string().nullish(),
      catchPayload: zod.object({}).passthrough().nullish(),
      createdAt: zod.string(),
    }),
  ),
});

/**
 * @summary Delete a session
 */
export const DeleteSessionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a session
 */
export const ListMessagesParams = zod.object({
  sessionId: zod.coerce.number(),
});

export const ListMessagesResponseItem = zod.object({
  id: zod.number(),
  sessionId: zod.number(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  intentType: zod.string().nullish(),
  catchPayload: zod.object({}).passthrough().nullish(),
  createdAt: zod.string(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary List decision ledger entries for a project
 */
export const ListEntriesParams = zod.object({
  projectId: zod.coerce.number(),
});

export const ListEntriesQueryParams = zod.object({
  status: zod.enum(["committed", "parked", "draft", "archived"]).optional(),
});

export const ListEntriesResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  sessionId: zod.number().nullish(),
  status: zod.enum(["committed", "parked", "draft", "archived"]),
  title: zod.string(),
  summary: zod.string().nullish(),
  severity: zod.enum(["blocker", "parked", "committed", "neutral"]),
  verb: zod.string().nullish(),
  isViolation: zod.boolean(),
  deviation: zod.boolean(),
  mode: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListEntriesResponse = zod.array(ListEntriesResponseItem);

/**
 * @summary Create a ledger entry
 */
export const CreateEntryParams = zod.object({
  projectId: zod.coerce.number(),
});

export const CreateEntryBody = zod.object({
  sessionId: zod.number().nullish(),
  status: zod.enum(["committed", "parked", "draft"]),
  title: zod.string(),
  summary: zod.string().nullish(),
  severity: zod.enum(["blocker", "parked", "committed", "neutral"]),
  verb: zod.string().nullish(),
  mode: zod.string().nullish(),
});

/**
 * @summary Update a ledger entry
 */
export const UpdateEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateEntryBody = zod.object({
  status: zod.enum(["committed", "parked", "draft", "archived"]).optional(),
  title: zod.string().optional(),
  summary: zod.string().nullish(),
  severity: zod.enum(["blocker", "parked", "committed", "neutral"]).optional(),
});

export const UpdateEntryResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  sessionId: zod.number().nullish(),
  status: zod.enum(["committed", "parked", "draft", "archived"]),
  title: zod.string(),
  summary: zod.string().nullish(),
  severity: zod.enum(["blocker", "parked", "committed", "neutral"]),
  verb: zod.string().nullish(),
  isViolation: zod.boolean(),
  deviation: zod.boolean(),
  mode: zod.string().nullish(),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a ledger entry
 */
export const DeleteEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Send a message and get an AI response
 */
export const SendMessageBody = zod.object({
  sessionId: zod.number(),
  projectId: zod.number(),
  message: zod.string(),
  mode: zod.string(),
  history: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
      }),
    )
    .optional(),
  entries: zod
    .array(
      zod.object({
        id: zod.number(),
        title: zod.string(),
        status: zod.string(),
      }),
    )
    .optional(),
});

export const SendMessageResponse = zod.object({
  content: zod.string(),
  intentType: zod.string().nullish(),
  catchPayload: zod.object({}).passthrough().nullish(),
  messageId: zod.number(),
});