# lib/api-zod/src/generated/types/chatHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatHistoryItemRole } from "./chatHistoryItemRole";

export interface ChatHistoryItem {
  role: ChatHistoryItemRole;
  content: string;
}