# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatHistoryItem } from "./chatHistoryItem";
import type { EntryContext } from "./entryContext";

export interface ChatRequest {
  sessionId: number;
  projectId: number;
  message: string;
  mode: string;
  history?: ChatHistoryItem[];
  entries?: EntryContext[];
}