# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatResponseCatchPayload } from "./chatResponseCatchPayload";

export interface ChatResponse {
  content: string;
  /** @nullable */
  intentType?: string | null;
  /** @nullable */
  catchPayload?: ChatResponseCatchPayload;
  messageId: number;
}