# lib/api-zod/src/generated/types/entryContext.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Atlas - Strategic Thinking Partner API
 * OpenAPI spec version: 0.1.0
 */

export interface EntryContext {
  id: number;
  title: string;
  status: string;
}